refactor: split device service by use case
